always take close wall, able test in MIC
The cbGetLines will append walls with the same direction. For example, two walls with different distance and same angle will append to the walls and angles list. There is a loop to check walls with similar angle, it delete the far wall and keep the close wall. So far so good, test in MIC. Need use start and end point in line message for the precision to detect the correct wall.